Moved list of exercises to it's own file
